+++ src/MunicipalityRegistry.Producer/Infrastructure/Program.cs
@@ -3,9 +3,13 @@ namespace MunicipalityRegistry.Producer.Infrastructure
     using Be.Vlaanderen.Basisregisters.Api;
     using Be.Vlaanderen.Basisregisters.Aws.DistributedMutex;
     using Microsoft.AspNetCore.Hosting;
+    using Microsoft.Extensions.Configuration;
 
     public class Program
     {
+        protected Program()
+        { }
+
         public static void Main(string[] args)
             => Run(new ProgramOptions
                 {