+++ src/MunicipalityRegistry.Projections.Wfs/Properties/AssemblyInfo.cs
@@ -1,0 +1,7 @@
+using System.Reflection;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyDescription("MunicipalityRegistry Wfs Projections")]
+
+[assembly: ComVisible(false)]
+[assembly: Guid("e27640d1-ebb4-4a84-852b-59b04c06745e")]