+++ src/MunicipalityRegistry.Projections.Wms/Properties/AssemblyInfo.cs
@@ -1,0 +1,7 @@
+using System.Reflection;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyDescription("MunicipalityRegistry Wms Projections")]
+
+[assembly: ComVisible(false)]
+[assembly: Guid("90b50fbe-ec3a-4fba-bba4-2c09e8210b26")]