+++ src/MunicipalityRegistry.Projector/Infrastructure/Startup.cs
@@ -24,6 +24,9 @@ namespace MunicipalityRegistry.Projector.Infrastructure
     using Modules;
     using MunicipalityRegistry.Projections.Extract;
     using MunicipalityRegistry.Projections.Legacy;
+    using MunicipalityRegistry.Projections.Wfs;
+    using MunicipalityRegistry.Projections.Wms;
+    using Swashbuckle.AspNetCore.Swagger;
 
     /// <summary>Represents the startup process for the application.</summary>
     public class Startup
@@ -111,6 +114,14 @@ namespace MunicipalityRegistry.Projector.Infrastructure
 
                                 health.AddDbContextCheck<LastChangedListContext>(
                                     $"dbcontext-{nameof(LastChangedListContext).ToLowerInvariant()}",
+                                    tags: new[] {DatabaseTag, "sql", "sqlserver"});
+
+                                health.AddDbContextCheck<WfsContext>(
+                                    $"dbcontext-{nameof(WfsContext).ToLowerInvariant()}",
+                                    tags: new[] {DatabaseTag, "sql", "sqlserver"});
+
+                                health.AddDbContextCheck<WmsContext>(
+                                    $"dbcontext-{nameof(WmsContext).ToLowerInvariant()}",
                                     tags: new[] {DatabaseTag, "sql", "sqlserver"});
                             }
                         }