+++ src/MunicipalityRegistry.Projector/Projections/ProjectionsController.cs
@@ -20,6 +20,8 @@ namespace MunicipalityRegistry.Projector.Projections
         {
             RegisterConnectionString(Schema.Legacy, configuration.GetConnectionString("LegacyProjections"));
             RegisterConnectionString(Schema.Extract, configuration.GetConnectionString("ExtractProjections"));
+            RegisterConnectionString(Schema.Wfs, configuration.GetConnectionString("WfsProjections"));
+            RegisterConnectionString(Schema.Wms, configuration.GetConnectionString("WmsProjections"));
         }
     }
 }